Add new mixin class for quotas.
