+++ releasenotes/notes/dcs_quota-2b8e740be0d4294a.yaml
@@ -1,4 +1,5 @@
 ---
 features:
   - |
-    Add DCS Quota.
+    Add DCS Quota list operation support.
+    Add new mixin class for quotas.
